[PAL/vm-common] Protect whole state of threads during context switch
The context switch consists of `sched_thread_uninterruptable()` /
`sched_thread()` / `sched_thread_wait()` and an asm-written
`save_context_and_restore_next()`. Previously, our code tried to protect
the whole context switch using a global `g_thread_list_lock` lock, but
it had bugs -- some parts of the state of the leaving thread and of the
to-be-scheduled thread were not correctly protected, and thus two CPUs
could race on this state and lead to e.g. stack corruptions.

By the thread state we mean all thread's metadata (fields like `state`),
place in the global thread list, CPU context (GPRs, XSAVE area) and
thread stack.

This commit extends the protected critical section to the very end of
the context switch which happens closer to the end of
`save_context_and_restore_next()`. Since the state of the leaving thread
is thus protected by the global lock, there is no need in `is_running`
atomic field anymore.

Additionally, each CPU needs a "scheduling stack" that is used in
`save_context_and_restore_next()` when the stack of the leaving thread
cannot be used anymore and we can't yet use the stack of the
to-be-scheduled thread.

Finally, a small bug with RFLAGS value of the leaving thread is fixed.
Previously our asm code in `save_context_and_restore_next()` could
accidentally change some bits in RFLAGS before it was saved.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>