[PAL/vm-common] Mark LAPIC's ICR as generally insecure
LAPIC has an ICR (Interrupt Command Register) to send IPIs. In Intel
TDX, this register is insecure, i.e., it must be accessed via the TDCALL
and is controlled by the untrusted host. Previously we didn't mark it
explicitly as insecure; this commit adds this marking.

Note: currently ICR is used only in `init_multicore_init_sipi_sipi()`
which is *not* executed in TDX case, so it was benign before. However,
next commits will introduce ICR usages for TDX (for "flush TLB" IPI
notifications among vCPUs).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>