[PAL/vm-common] Explicitly allocate memory on file mmap operation
Previuosly, file mmap callback did not call `memory_alloc()` to request
a memory region to be allocated into which the file contents will be
copied. This was a bug that worked before all pages are currently marked
as Present and RWX, so `memory_alloc()` is currently a no-op (other than
zeroing out the memory).

This bug was uncovered when implementing proper page permissions for
VM-based PALs. This commit adds a required call to `memory_alloc()`.
Note that this leads to a slight performance overhead, as we now
redundantly memset the newly-allocated memory region to all-zeros (and
then immediately overwrite with file contents). However, the file-mmap
operation is sufficiently rare, and correct memory management is more
important than a small perf overhead.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>