[PAL/vm-common] Get memory regions from VMM using "etc/e820" fw_cfg file
Previously on the VM PAL, we hard-coded a single memory region that is
RAM (i.e., not reserved) via `FW_CFG_RAM_SIZE`. This is a dangerous
assumption which may break in the future. This commit uses a proper
source of E820 info: the `etc/e820` fw_cfg file.

As a side effect, `find_fw_cfg_selector()` function was slightly
modified to also return the size of the object. All callers of this
function now use this returned size.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>